Silence the MSVC and gcc warnings the Mac build cannot see
All pre-existing, from build run 30726164386. The Mac/clang build was and
stays at zero warnings.

MSVC C4996 (deprecated CRT calls) - fopen replaced with std::ofstream /
ifstream in core/synfile.cpp, core/bmpfile.cpp, cli/kgfax-decode.cpp and
gui/main.cpp's --test-scan dump, the way core/settings.cpp and the two
test writers already went. localtime gets a small portable wrapper
(localtime_s on MSVC, localtime_r elsewhere), which also fixes a real
thread-safety wart since localtime returns a shared buffer.

MSVC C4244 - gui/progressdialog.cpp passed a double to Fl_Progress::value,
which takes float. Explicit cast.

gcc -Wformat-truncation in gui/main.cpp's set_title: the suffix is
caller-built (file paths, decode status) and could be truncated into the
160-byte buffer. Built with std::string instead, so it cannot truncate at
all.

gcc -Wmaybe-uninitialized pointed at core/syncscan.h:29, which is where
FaxImage is declared, and there was a genuine uninitialized field behind
it: syn_read set lines_locked, lines_coasted and relocks but never
lines_corrected. Now initialised. Confirmed real - the pre-change binary
reports "corrected 1" on a freshly read .syn where the fixed one reports
0.

Byte-for-byte verified against the pre-change code on a synthetic image:
.syn and .bmp output are both identical, so the stream rewrites changed
no format. 11/11 ctest green, clean build, no warnings on macOS.

+++ gui/progressdialog.cpp
@@ -61,6 +61,6 @@ void ProgressScope::update(double fraction)
     ProgressWin *pw = (ProgressWin *)win;
     if (fraction < 0) fraction = 0;
     if (fraction > 1) fraction = 1;
-    pw->bar->value(fraction);
+    pw->bar->value((float)fraction);   /* Fl_Progress takes float (C4244) */
     Fl::check();   /* keep the window responsive + repainted */
 }
